mess around with an alternate parser
parser conversion notes, I had to modify the comment regex,
to get it working with this lexer.

As author of the thing consuming nimbleparse.toml,
https://github.com/ratmice/nimbleparse.toml

I was curious to mess around with existing language projects,
To see if there were any pain points regarding project structure,
and see if it was fast enough for interactive use in projects with more
tests.

At least one thing is apparent,
nimbleparse.toml really needs a recursive directory walker...